game of bet
Write a Python program that simulates a simple slot machine in which three numbers
between 0 and 9 are randomly selected and printed side by side. The user has an initial sum
of $100. The user enters his/her bet (make sure it is a valid bet, i.e., the bet is positive and
less than or equal to the sum). If the three numbers are the same, the user hits the Jackpot, this
means that he wins ten times his/her bet. If only two of the three numbers are the same the
user wins four times his/her bet. After each round, the program outputs the sum that the user
has and asks the user if he/she wish to continue playing. If yes the program repeats, else, the
program displays the total winnings or losses of the user and then ends.